Reorganize files and load HTML at compile time
Improvised the readability of the project
HTML update means files no longer need to be present at runtime to work
